% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_muac.R
\name{add_muac}
\alias{add_muac}
\title{add_muac}
\usage{
add_muac(
  .dataset,
  nut_muac_cm = "nut_muac_cm",
  edema_confirm = "nut_edema_confirm",
  child_age_months = "child_age_months",
  child_sex = "child_sex",
  value_male_sex = "m",
  value_edema_confirm = "yes"
)
}
\arguments{
\item{.dataset}{Child Nutrition Loop Dataset}

\item{nut_muac_cm}{the name of the variable that indicates the MUAC measurement by CM.
By default it is nut_muac_cm}

\item{edema_confirm}{the name of the variable that indicates that edema is confirmed
By default it is nut_edema_confirm}

\item{child_age_months}{the name of the variable that indicates the age of child per month.
By default it is child_age_months}

\item{child_sex}{the name of the variable that indicates the sex of the child
By default it is child_sex}

\item{value_male_sex}{the value of the choice "male" of the sex indicator
By default it is m}

\item{value_edema_confirm}{the value of the choice "yes" of the nut_edema_confirm indicator
By default it is "yes"}
}
\value{
the dataset with the sam_muac, mam_muac, and gam_muac calcualted
}
\description{
add_muac
}
\examples{
df1 <- data.frame(
uuid = c("uuid_1","uuid_2"),
nut_muac_cm = c("12.5","10"),
child_sex = c("m","f"),
child_age_months = c("14","54"),
nut_edema_confirm = c("yes",NA))

add_muac(.dataset = df1)
}
