% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_persontime.R
\name{add_persontime}
\alias{add_persontime}
\title{Add Person Time}
\usage{
add_persontime(df_mortality, smart = FALSE)
}
\arguments{
\item{df_mortality}{long dataframe containing (roster/left/died dfs)}

\item{smart}{boolean to check if calculation should follow smart methodology or IMPACT}
}
\value{
Long dataframe with person time added
}
\description{
Add Person Time
}
\examples{
\dontrun{
  add_persontime(df_mortality, smart = F)
}
}
