% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plausibility.R
\name{calculate_plausibility}
\alias{calculate_plausibility}
\title{Calculate the plausibility for IYCF/MORT/NUT/FSL}
\usage{
calculate_plausibility(.dataset)
}
\arguments{
\item{.dataset}{the output of create_x_plaus}
}
\value{
dataframe with plausibility columns calculated
}
\description{
Calculate the plausibility for IYCF/MORT/NUT/FSL
}
\examples{
\dontrun{calculate_plausibility(df)}
}
