% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_anthro_flags.R
\name{check_anthro_flags}
\alias{check_anthro_flags}
\title{check_anthro_flags}
\usage{
check_anthro_flags(
  .dataset,
  nut_muac_cm = "nut_muac_cm",
  edema_confirm = "nut_edema_confirm",
  value_edema_confirm = "yes",
  uuid = "uuid",
  loop_index = NULL
)
}
\arguments{
\item{.dataset}{the raw child nutrition loop dataset with all add_muac and add_mfaz called}

\item{nut_muac_cm}{the name of the variable that indicates the MUAC measurement by CM.
By default it is nut_muac_cm}

\item{edema_confirm}{the name of the variable that indicates that edema is confirmed
By default it is nut_edema_confirm}

\item{value_edema_confirm}{the value of the choice "yes" of the nut_edema_confirm indicator
By default it is "yes"}

\item{uuid}{uuid variable}

\item{loop_index}{unique identifier for each individual in the child nutrition loop.
By default, loop_index}
}
\value{
a dataframe that includes all the logical flags related to nutrition
This includes:
- flag_sd_mfaz
- mfaz_who_flag
- mfaz_noflag
- mean_mfaz_noflag
- sd_mfaz_noflag
- global_mfaz_noflag
- moderate_mfaz_noflag
- severe_mfaz_noflag
- flag_extreme_muac
- muac_noflag
- gam_muac_noflag
- mam_muac_noflag
- sam_muac_noflag
- flag_severe_hhs
- flag_edema_pitting
}
\description{
check_anthro_flags
}
\examples{
\dontrun{check_anthro_flags(df)}
}
