% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_iycf_flags.R
\name{check_iycf_flags}
\alias{check_iycf_flags}
\title{check_iycf_flags}
\usage{
check_iycf_flags(
  .dataset,
  age_months = "age_months",
  iycf_4 = "iycf_4",
  iycf_6a = "iycf_6a",
  iycf_6b = "iycf_6b",
  iycf_6c = "iycf_6c",
  iycf_6d = "iycf_6d",
  iycf_6e = "iycf_6e",
  iycf_6f = "iycf_6f",
  iycf_6g = "iycf_6g",
  iycf_6h = "iycf_6h",
  iycf_6i = "iycf_6i",
  iycf_6j = "iycf_6j",
  iycf_7a = "iycf_7a",
  iycf_7b = "iycf_7b",
  iycf_7c = "iycf_7c",
  iycf_7d = "iycf_7d",
  iycf_7e = "iycf_7e",
  iycf_7f = "iycf_7f",
  iycf_7g = "iycf_7g",
  iycf_7h = "iycf_7h",
  iycf_7i = "iycf_7i",
  iycf_7j = "iycf_7j",
  iycf_7k = "iycf_7k",
  iycf_7l = "iycf_7l",
  iycf_7m = "iycf_7m",
  iycf_7n = "iycf_7n",
  iycf_7o = "iycf_7o",
  iycf_7p = "iycf_7p",
  iycf_7q = "iycf_7q",
  iycf_7r = "iycf_7r",
  iycf_8 = "iycf_8",
  iycf_6b_num = "iycf_6b_num",
  iycf_6c_num = "iycf_6c_num",
  iycf_6d_num = "iycf_6d_num",
  yes_value = "yes",
  uuid = "uuid"
)
}
\arguments{
\item{.dataset}{the raw child nutrition loop dataset with all iycf indicators}

\item{age_months}{the name of the variable that indicates the age of child per month.}

\item{iycf_4}{the name of the variable that indicates if the child was breastfed yesterday
By default "iycf_4"}

\item{iycf_6a}{the name of the variable that indicates if the child
had water yesterday. By default "iycf_6a"}

\item{iycf_6b}{the name of the variable that indicates if the child
had infant formula yesterday. By default "iycf_6b"}

\item{iycf_6c}{the name of the variable that indicates if the child
had milk from animal yesterday. By default "iycf_6c"}

\item{iycf_6d}{the name of the variable that indicates if the child
had yogurt drink yesterday. By default "iycf_6d"}

\item{iycf_6e}{the name of the variable that indicates if the child
had chocolate flavored drink yesterday. By default "iycf_6e"}

\item{iycf_6f}{the name of the variable that indicates if the child
had fruit juice or fruit flavored drink yesterday. By default "iycf_6f"}

\item{iycf_6g}{the name of the variable that indicates if the child
had sodas/malt/sports/energy drink yesterday. By default "iycf_6g"}

\item{iycf_6h}{the name of the variable that indicates if the child
had tea/coffee/herbal drink yesterday. By default "iycf_6h"}

\item{iycf_6i}{the name of the variable that indicates if the child
had clear broth/soup yesterday. By default "iycf_6i"}

\item{iycf_6j}{the name of the variable that indicates if the child
had other liquids yesterday. By default "iycf_6i"}

\item{iycf_7a}{the name of the variable that indicates if the child
had yogurt as food yesterday. By default "iycf_7a"}

\item{iycf_7b}{the name of the variable that indicates if the child
had porridge/bread/rice/noodles/pasta yesterday. By default "iycf_7b"}

\item{iycf_7c}{the name of the variable that indicates if the child
had pumpkin/carrots/sweet red peppers/squash/sweet potato
yesterday. By default "iycf_7c"}

\item{iycf_7d}{the name of the variable that indicates if the child
had plantains/white potato/yams/manioc/cassava
yesterday. By default "iycf_7d"}

\item{iycf_7e}{the name of the variable that indicates if the child
had dark green leafy vegetables yesterday. By default "iycf_7e"}

\item{iycf_7f}{the name of the variable that indicates if the child
had other vegetables yesterday. By default "iycf_7f"}

\item{iycf_7g}{the name of the variable that indicates if the child
had ripe mangoes/ripe papayas yesterday. By default "iycf_7g"}

\item{iycf_7h}{the name of the variable that indicates if the child
had other fruits yesterday. By default "iycf_7h"}

\item{iycf_7i}{the name of the variable that indicates if the child
had liver/kidney/heart yesterday. By default "iycf_7i"}

\item{iycf_7j}{the name of the variable that indicates if the child
had sausage/hot dogs/ham/bacon/salami/canned meat
yesterday. By default "iycf_7j"}

\item{iycf_7k}{the name of the variable that indicates if the child
had other meat/beef/pork/lamb/goat/chicken/duck
yesterday. By default "iycf_7k"}

\item{iycf_7l}{the name of the variable that indicates if the child
had eggs yesterday. By default "iycf_7l"}

\item{iycf_7m}{the name of the variable that indicates if the child
had fresh/dried/shell fish yesterday. By default "iycf_7m"}

\item{iycf_7n}{the name of the variable that indicates if the child
had beans/peas/lentils/nuts/seeds yesterday. By default "iycf_7n"}

\item{iycf_7o}{the name of the variable that indicates if the child
had hard/soft cheese yesterday. By default "iycf_7o"}

\item{iycf_7p}{the name of the variable that indicates if the child
had sweet foods yesterday. By default "iycf_7p"}

\item{iycf_7q}{the name of the variable that indicates if the child
had chips/crisps/puffs/french fries yesterday. By default "iycf_7q"}

\item{iycf_7r}{the name of the variable that indicates if the child
had other solid food yesterday. By default "iycf_7q"}

\item{iycf_8}{the name of the variable that indicates if the meal frequency the
child had yesterday. By default "iycf_8"}

\item{iycf_6b_num}{the name of the variable that indicates how many times the child
had infant formula yesterday. By default "iycf_6b_num"}

\item{iycf_6c_num}{the name of the variable that indicates how many times the child
had milk from animal yesterday. By default "iycf_6c_num"}

\item{iycf_6d_num}{the name of the variable that indicates how many times the child
had yogurt drink yesterday. By default "iycf_6d_num"}

\item{yes_value}{the value of the choice "yes" to all the liquid/food categories}

\item{uuid}{uuid variable}
}
\value{
a dataframe that includes all the added flags related to iycf.
This includes:
- flag_yes_foods: all foods selected were yes
- flag_yes_liquids: all liquids were yes
- flag_no_anything: no foods or liquids at all
- flag_no_foods: no food at all and mealfreq is more than 0 and age_months more than 5
- flag_all_foods_no_meal: all food given but mealfreq is equal to 0
- flag_some_foods_no_meal: some food given but mealfreq is equal to 0
- flag_high_mdd_low_mmf: High Minimum Dietary Diversity and low mealfreq
- flag_under6_nobf_nomilk: child under 6 with no milk or breastfeeding
- flag_meats_nostaples: child given meat but no staples
}
\description{
check_iycf_flags
}
\examples{
\dontrun{check_iycf_flags(df1)}
}
