% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mortality_flags.R
\name{check_mortality_flags}
\alias{check_mortality_flags}
\title{check_mortality_flags}
\usage{
check_mortality_flags(
  df_mortality,
  cause_death = "cause_death",
  cause_death_f = c("post_partum", "during_pregnancy", "during_delivery")
)
}
\arguments{
\item{df_mortality}{output dataframe long mortality from create_mortality_long_df}

\item{cause_death_f}{vector list of the cause of death options related to female}
}
\value{
df_mortality with two extra flag columns:
-flag_multiple_death
-flag_cause_death
}
\description{
check_mortality_flags
}
\examples{
\dontrun{
  check_mortality_flags(df_mortality,
  cause_death_f = c("post_partum","during_pregnancy","during_delivery"))
}
}
