% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fsl_analysis_template.R
\name{create_fsl_analysis_template}
\alias{create_fsl_analysis_template}
\title{Create a project folder for fsl analysis}
\usage{
create_fsl_analysis_template(folder_path, ...)
}
\arguments{
\item{folder_path}{Path for the new folder}

\item{...}{Extra information collected from the RStudio wizard}
}
\value{
Folder with a copy of the fsl_analysis
}
\description{
Create a project folder for fsl analysis
}
\examples{
\dontrun{
create_fsl_analysis_template("path/to/folder")
}
}
