% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_iycf_cleaning_template.R
\name{create_iycf_cleaning_template}
\alias{create_iycf_cleaning_template}
\title{Create a project folder for iycf cleaning}
\usage{
create_iycf_cleaning_template(folder_path, ...)
}
\arguments{
\item{folder_path}{Path for the new folder}

\item{...}{Extra information collected from the RStudio wizard}
}
\value{
Folder with a copy of the iycf_cleaning
}
\description{
Create a project folder for iycf cleaning
}
\examples{
\dontrun{
create_iycf_cleaning_template("path/to/folder")
}
}
