% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_iycf_plaus.R
\name{create_iycf_plaus}
\alias{create_iycf_plaus}
\title{create_iycf_plaus}
\usage{
create_iycf_plaus(
  df_iycf,
  age_months = NULL,
  sex = NULL,
  iycf_8 = NULL,
  iycf_caregiver = NULL,
  yes_value_caregiver = NULL,
  no_value_caregiver = NULL,
  exp_prevalence_mad = NULL,
  exp_sex_ratio = NULL,
  exp_ratio_under6m_6to23m = NULL,
  grouping = NULL,
  uuid = "uuid",
  short_report = FALSE,
  file_path = NULL
)
}
\arguments{
\item{df_iycf}{dataframe output of the check_iycf_flag function}

\item{age_months}{the name of the variable that indicates the age in month of the child}

\item{sex}{the name of the variable that indicates the sex of the child}

\item{iycf_8}{the name of the variable that indicates if the meal frequency the
child had yesterday. By default "iycf_8"}

\item{iycf_caregiver}{the name of the variable that indicates if the caregiver
of the child is present. By default NULL}

\item{yes_value_caregiver}{the value of the choice "yes" to all the caregiver column}

\item{no_value_caregiver}{the value of the choice "no" to all the caregiver column}

\item{exp_prevalence_mad}{Expected prevalence for Minimum Acceptable Diet (MAD)
By default: 0.3:0.7}

\item{exp_sex_ratio}{Expected sex ratio. By default: 1:1}

\item{exp_ratio_under6m_6to23m}{Expected age ratio between children under 6 month,
and children between 6 and 23 months. By default: 1:4}

\item{grouping}{the name of the variable that indicates the grouping variable - usually "enumerator"}

\item{uuid}{uuid variable}

\item{short_report}{Inputs a boolean value TRUE or FALSE to return just key variables. If FALSE,
returns a dataframe of all the variables calculated.}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}
}
\value{
a dataframe with all IYCF related plausibility columns
}
\description{
create_iycf_plaus
}
\examples{
\dontrun{
  create_iycf_plaus(df_iycf)
}
}
