% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mort_quality_report_template.R
\name{create_mort_quality_report_template}
\alias{create_mort_quality_report_template}
\title{Create a project folder for mort quality report}
\usage{
create_mort_quality_report_template(folder_path, ...)
}
\arguments{
\item{folder_path}{Path for the new folder}

\item{...}{Extra information collected from the RStudio wizard}
}
\value{
Folder with a copy of the mort_quality_report
}
\description{
Create a project folder for mort quality report
}
\examples{
\dontrun{
create_mort_quality_report_template("path/to/folder")
}
}
