% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_muac_plaus.R
\name{create_muac_plaus}
\alias{create_muac_plaus}
\title{create_muac_plaus}
\usage{
create_muac_plaus(df, grouping = NULL, file_path = NULL, short_report = NULL)
}
\arguments{
\item{df}{raw/clean data with all calculated
fcs/rcsi/hhs/hdds/fcm/fclcm add_x indicators}

\item{grouping}{the name of the variable that indicates the grouping variable - usually "enumerator"}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}

\item{short_report}{Inputs a boolean value TRUE or FALSE to return just key variables. If FALSE,
returns a dataframe of all the variables calculated.}
}
\value{
a dataframe with all fsl related plausibility columns
}
\description{
create_muac_plaus
}
\examples{
\dontrun{create_muac_plaus(df)}
}
