% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ph_integrated_tables_template.R
\name{create_ph_integrated_tables_template}
\alias{create_ph_integrated_tables_template}
\title{Create a project folder for ph integrated tables}
\usage{
create_ph_integrated_tables_template(folder_path, ...)
}
\arguments{
\item{folder_path}{Path for the new folder}

\item{...}{Extra information collected from the RStudio wizard}
}
\value{
Folder with a copy of the ph_integrated_tables
}
\description{
Create a project folder for ph integrated tables
}
\examples{
\dontrun{
create_ph_integrated_tables_template("path/to/folder")
}
}
