% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plaus_table_fsl.R
\name{create_plaus_table_fsl}
\alias{create_plaus_table_fsl}
\title{Create the plausibility table for FSL}
\usage{
create_plaus_table_fsl(.flextable)
}
\arguments{
\item{.flextable}{A flextable format of the return value of create_fsl_plaus}
}
\value{
a Flextable of the FSL plausibility report
}
\description{
Create the plausibility table for FSL
}
\examples{
\dontrun{
  create_plaus_table_fsl(fsl_flextable)
}
}
