% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plaus_table_iycf.R
\name{create_plaus_table_iycf}
\alias{create_plaus_table_iycf}
\title{Create the plausibility table for IYCF}
\usage{
create_plaus_table_iycf(.flextable)
}
\arguments{
\item{.flextable}{A flextable format of the return value of create_iycf_plaus}
}
\value{
a Flextable of the IYCF plausibility report
}
\description{
Create the plausibility table for IYCF
}
\examples{
\dontrun{
  create_plaus_table_iycf(iycf_flextable)
}
}
