% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plaus_table_nut_mort.R
\name{create_plaus_table_nut_mort}
\alias{create_plaus_table_nut_mort}
\title{Create the plausibility table for Nutrition and Mortality}
\usage{
create_plaus_table_nut_mort(.flextable)
}
\arguments{
\item{.flextable}{A flextable format of the return value of create_mortality_plaus
and create_nut_plaus}
}
\value{
a Flextable of the Mortality and Nut plausibility report
}
\description{
Create the plausibility table for Nutrition and Mortality
}
\examples{
\dontrun{
  create_plaus_table_nut_mort(mort_or_nut_flextable)
}
}
