% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.label.R
\name{get.label}
\alias{get.label}
\title{Find the label of a variable}
\usage{
get.label(variable = NULL, tool.survey = NULL, language = "English")
}
\arguments{
\item{variable}{This is the name of the header from raw data. By default NULL}

\item{tool.survey}{This is the tool.survey dataset. By default NULL}

\item{language}{This is the language of the label:: in the tool. By default English}
}
\value{
return label will be the one for the base question itself
   (e.g. if variable == "water_source", the result is the label of "What are
   the water sources in your household")
}
\description{
Find the label of a variable
}
\examples{
\dontrun{
label <- get.label("water_source", tool.survey = tool.survey)
}
}
