% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{impactR4PHU_choice_template}
\alias{impactR4PHU_choice_template}
\title{MSNA template choices tool (example)}
\usage{
impactR4PHU_choice_template
}
\description{
MSNA template choices tool (example)
}
\keyword{datasets}
