% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{impactR4PHU_data_health_template}
\alias{impactR4PHU_data_health_template}
\title{MSNA template health dataset (example)}
\usage{
impactR4PHU_data_health_template
}
\description{
MSNA template health dataset (example)
}
\keyword{datasets}
