% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{impactR4PHU_data_mort_template}
\alias{impactR4PHU_data_mort_template}
\title{MSNA template mortality dataset (example)}
\usage{
impactR4PHU_data_mort_template

impactR4PHU_data_mort_template
}
\description{
MSNA template mortality dataset (example)

MSNA template mortality dataset (example)
}
\keyword{datasets}
