% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{impactR4PHU_data_roster_template}
\alias{impactR4PHU_data_roster_template}
\title{IPHRA template roster dataset (example)}
\usage{
impactR4PHU_data_roster_template
}
\description{
IPHRA template roster dataset (example)
}
\keyword{datasets}
