% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{impactR4PHU_data_template}
\alias{impactR4PHU_data_template}
\title{MSNA template dataset (example)}
\usage{
impactR4PHU_data_template
}
\description{
Survey data
}
\details{
These data sets include HH data used for the readme
}
\examples{
impactR4PHU_data_template
}
\keyword{datasets}
