% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_set_documentation.R
\docType{data}
\name{impactR4PHU_MSNA_template_data}
\alias{impactR4PHU_MSNA_template_data}
\title{MSNA template main dataset (example)}
\usage{
impactR4PHU_MSNA_template_data
}
\description{
MSNA template main dataset (example)
}
\keyword{datasets}
