% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.label_colname.R
\name{load.label_colname}
\alias{load.label_colname}
\title{get the label::x colname from the tool.survey}
\usage{
load.label_colname(tool.survey = NULL, language = "English")
}
\arguments{
\item{tool.survey}{This is the tool.survey dataset. By default NULL}

\item{language}{This is the language desired to be selected. By default "English"}
}
\value{
return the name of the label column (ex: label::English)
}
\description{
get the label::x colname from the tool.survey
}
\examples{
\dontrun{
label_colname <- load.label_colname(tool.survey, language = "English")
}
}
