% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_distribution.R
\name{plot_age_distribution}
\alias{plot_age_distribution}
\title{Plot Age Distribution}
\usage{
plot_age_distribution(
  .dataset,
  by_group = NULL,
  year_or_month = NULL,
  age_years = "age_years",
  age_months = "age_months",
  min_age = NULL,
  max_age = NULL,
  breaks = NULL,
  file_path = NULL,
  wdth = 5,
  hght = 5,
  title_name = NULL
)
}
\arguments{
\item{.dataset}{HH_roster Data with Age and Sex of individuals}

\item{by_group}{Disaggregated variable}

\item{year_or_month}{Month or Year Plot. "year" for Year, "month" for Month.
By deault it will calculate year.}

\item{age_years}{the variable name in the Data indicating the year age of the individual
By default: "age_years"}

\item{age_months}{the variable name in the Data indicating the month age of the individual
By default: "age_months"}

\item{min_age}{Minimum Age cutoff. By default 0 for year and month.}

\item{max_age}{Maximum Age cutoff. By default 5 for year and 59 for month.}

\item{breaks}{Breaks of the Bins. By default 1 for year and 12 for month.}

\item{file_path}{File path for saving the plot in a directory}

\item{wdth}{width of the plot. By default: 5}

\item{hght}{height of the plot. By default: 5}

\item{title_name}{Title of the plot}
}
\value{
an age per year or month distribution plot
}
\description{
Plot Age Distribution
}
\examples{
\dontrun{
  plot_age_distribution(hh_roster)
}
}
