% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlogram.R
\name{plot_correlogram}
\alias{plot_correlogram}
\title{Plot Correlogram}
\usage{
plot_correlogram(
  .dataset,
  numeric_cols = c("fsl_fcs_score", "fsl_rcsi_score", "fsl_hhs_score"),
  file_path = NULL,
  wdth = 5,
  hght = 5,
  title_name = NULL
)
}
\arguments{
\item{.dataset}{output of the create_fsl_flags functions}

\item{numeric_cols}{a vector of the fsl indicator scores.
By default: c("fsl_fcs_score",  "fsl_rcsi_score",  "fsl_hhs_score")}

\item{file_path}{File path for saving the plot in a directory}

\item{wdth}{width of the plot. By default: 5}

\item{hght}{height of the plot. By default: 5}

\item{title_name}{Title of the plot}
}
\value{
a Correlogram plot
}
\description{
Plot Correlogram
}
\examples{
\dontrun{
  plot_correlogram(.dataset)
}
}
