% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridge_distribution.R
\name{plot_ridge_distribution}
\alias{plot_ridge_distribution}
\title{Plot Ridge Distribution}
\usage{
plot_ridge_distribution(
  .dataset,
  numeric_cols = NULL,
  name_groups = NULL,
  name_units = NULL,
  grouping = NULL,
  file_path = NULL,
  wdth = 5,
  hght = 5,
  title_name = NULL
)
}
\arguments{
\item{.dataset}{output of the create_fsl_flags functions}

\item{numeric_cols}{a vector of the same fsl indicator score columns
By default: NULL.}

\item{name_groups}{Name of the groups. By default: "Groups"}

\item{name_units}{Name of the units. By default: "Units"}

\item{grouping}{Variable name from the output create_fsl_flags for grouping}

\item{file_path}{File path for saving the plot in a directory}

\item{wdth}{width of the plot. By default: 5}

\item{hght}{height of the plot. By default: 5}

\item{title_name}{Title of the plot}
}
\value{
A Ridge Plot with the distribution of values
}
\description{
Plot Ridge Distribution
}
\examples{
\dontrun{
  plot_ridge_distribution(df)
}
}
