% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_mortality.R
\name{reformat_mortality}
\alias{reformat_mortality}
\title{Reformat Mortality Data}
\usage{
reformat_mortality(df)
}
\arguments{
\item{df}{long dataframe containing (roster/left/died dfs)}
}
\value{
dataframe with recoded values for:
-sex
-date_dc
-date_recall
-date_join
-date_birth
-date_left
-date_death
-join
-left
-birth
-death
-cause_death
-location_death
}
\description{
Reformat Mortality Data
}
\examples{
\dontrun{
  reformat_mortality(df_mortality)
}
}
