% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fsl_plaus_html_report.R
\name{run_fsl_plaus_html_report}
\alias{run_fsl_plaus_html_report}
\title{run_fsl_plaus_html_report}
\usage{
run_fsl_plaus_html_report(
  .dataset = NULL,
  uuid_var = "_uuid",
  group_var = "enum_id",
  output_file = "fsl_quality_report.pdf",
  output_dir = "reports"
)
}
\arguments{
\item{.dataset}{Main Dataset}

\item{uuid_var}{the name of the variable that indicates the uuid column}

\item{group_var}{the name of the variable that indicates the grouping column, usually enumerator ID column}

\item{output_file}{the name of the output file}

\item{output_dir}{the directory to export the output file}
}
\value{
an HTML rmarkdown file with FSL quality summaries
}
\description{
run_fsl_plaus_html_report
}
