% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_muac_plaus_html_report.R
\name{run_muac_plaus_html_report}
\alias{run_muac_plaus_html_report}
\title{run_muac_plaus_html_report}
\usage{
run_muac_plaus_html_report(
  .dataset = NULL,
  uuid_var = "_uuid",
  yes_no_team = "no",
  team_var = "team",
  group_var = "enum_id",
  output_file = "muac_quality_report.html",
  output_dir = "reports"
)
}
\arguments{
\item{.dataset}{Main Dataset}

\item{uuid_var}{the name of the variable that indicates the uuid column}

\item{team_var}{the name of the variable that indicates the team column, if it exists}

\item{group_var}{the name of the variable that indicates the grouping column, usually enumerator ID column}

\item{output_file}{the name of the output file}

\item{output_dir}{the directory to export the output file}
}
\value{
an HTML rmarkdown file with MUAC quality summaries
}
\description{
run_muac_plaus_html_report
}
