% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subch.R
\name{subch}
\alias{subch}
\title{subch}
\usage{
subch(g, fig_height = 7, fig_width = 5)
}
\arguments{
\item{g}{figure to be included as a subchunk in rmarkdown}

\item{fig_height}{height of the figure}

\item{fig_width}{width of the figure}
}
\value{
a figure to be included as a subchunk in rmarkdown
}
\description{
subch
}
