% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wash_flags.R
\name{check_wash_flags}
\alias{check_wash_flags}
\title{check_wash_flags}
\usage{
check_wash_flags(
  .dataset,
  data_container_loop = NULL,
  container_type = "wash_container_type",
  container_litre_other = "wash_container_litre_other",
  container_journey_collection = "wash_container_journey_collection",
  num_containers = "wash_num_containers",
  water_source = "wash_water_source",
  different_water_sources = NULL,
  water_collect_time = "wash_water_collect_time",
  choice_inside_compound = "inside_compound",
  num_hh = "num_hh",
  grouping = NULL,
  uuid = "uuid"
)
}
\arguments{
\item{.dataset}{the raw dataset}

\item{data_container_loop}{loop_dataset for containers. By default NULL}

\item{container_type}{the name of the variable that indicates the type of the container including in the choice
the amount per liter (ex: choicex_20l)}

\item{container_litre_other}{the name of the variable that indicates the amount per liter for other containers}

\item{container_journey_collection}{the name of the variable that indicates the amount of journeys done}

\item{num_containers}{the name of the variable that indicates the number of containers per HH}

\item{water_source}{the name of the variable that indicates the main water source}

\item{different_water_sources}{the name of the variable that indicates the other water source (Select_multiple up to 2)
by default NULL, but better to call it if you collect this indicator.}

\item{water_collect_time}{the name of the variable that indicates the collection time}

\item{choice_inside_compound}{the name of the choice inside the water_collection_time variable that indicate
that the collection time is none because water source is inside the compound. By default: (inside_compound)}

\item{num_hh}{the name of the variable that indicates the number of people per HH}

\item{grouping}{the name of the variable that indicates the grouping variable - usually "enumerator"}

\item{uuid}{uuid variable}
}
\value{
a dataframe that includes all the logical flags related to WASH
This includes:
If loop containers available:
- litre_per_day_per_person
- litre_z_score
- flag_sd_litre
- flag_low_litre
- flag_high_litre
- flag_high_container
- flag_no_container
- flag_not_immediate
If loop containers not available:
- flag_not_immediate
}
\description{
check_wash_flags
}
\examples{
df <- data.frame(
  num_hh = c("3","4"),
  wash_num_containers = c("2","3"),
  enumerator = c("1","2"),
  wash_water_source = c("asdasd", "piped_dwelling"),
  wash_water_collect_time = c("inside_compound", "aasdasd"),
  uuid = c("uuid_1","uuid_2")
)

df_containers <- data.frame(
  uuid = c("uuid_1","uuid_1","uuid_2","uuid_2","uuid_2"),
  wash_container_type = c("bucket_20l", "bucket_14l", "jerry_can_10l",
                          "collapsible_jerry_can_5l" ,"other"),
  wash_container_litre_other = c(NA,NA,NA,NA,"30"),
  wash_container_journey_collection = c("1","2", "1","1","1")
)

check_wash_flags(
  .dataset = df,
  data_container_loop = df_containers,
  grouping = "enumerator"
)
}
