% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fsl_plaus.R
\name{create_fsl_plaus}
\alias{create_fsl_plaus}
\title{create_fsl_plaus}
\usage{
create_fsl_plaus(
  .dataset,
  grouping = NULL,
  uuid = "uuid",
  short_report = FALSE,
  file_path = NULL
)
}
\arguments{
\item{.dataset}{raw/clean data with all calculated
fcs/rcsi/hhs/hdds/fcm/fclcm add_x indicators}

\item{grouping}{the name of the variable that indicates the grouping variable - usually "enumerator"}

\item{uuid}{uuid variable}

\item{short_report}{Inputs a boolean value TRUE or FALSE to return just key variables. If FALSE,
returns a dataframe of all the variables calculated.}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}
}
\value{
a dataframe with all fsl related plausibility columns
}
\description{
create_fsl_plaus
}
\examples{
\dontrun{create_fsl_plaus(df)}
}
