% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mortality_plaus.R
\name{create_mortality_plaus}
\alias{create_mortality_plaus}
\title{create_mortality_plaus}
\usage{
create_mortality_plaus(
  df_mortality,
  exp_sex_ratio = NULL,
  exp_ratio_0_4 = NULL,
  exp_ratio_2_5 = NULL,
  exp_ratio_5_10 = NULL,
  exp_hh_size = NULL,
  grouping = NULL,
  uuid = "uuid",
  short_report = FALSE,
  file_path = NULL
)
}
\arguments{
\item{df_mortality}{output dataframe long mortality from create_mortality_long_df}

\item{exp_sex_ratio}{## to continue here}

\item{exp_ratio_0_4}{## to continue here}

\item{exp_ratio_2_5}{## to continue here}

\item{exp_ratio_5_10}{## to continue here}

\item{exp_hh_size}{## to continue here}

\item{grouping}{the name of the variable that indicates the grouping variable - usually "enumerator"}

\item{uuid}{uuid variable}

\item{short_report}{Inputs a boolean value TRUE or FALSE to return just key variables. If FALSE,
returns a dataframe of all the variables calculated.}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}
}
\value{
## to continue here
}
\description{
create_mortality_plaus
}
\examples{
\dontrun{
create_mortality_plaus(df_mortality)
}
}
