% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_muac_plaus.R
\name{create_muac_plaus}
\alias{create_muac_plaus}
\title{create_muac_plaus}
\usage{
create_muac_plaus(df, grouping = NULL, file_path = NULL, short_report = NULL)
}
\arguments{
\item{grouping}{the name of the variable that indicates the grouping variable - usually "enumerator"}

\item{file_path}{Inputs an optional character value specifying the file location to save a copy
of the results.}

\item{short_report}{Inputs a boolean value TRUE or FALSE to return just key variables. If FALSE,
returns a dataframe of all the variables calculated.}

\item{.dataset}{raw/clean data with all calculated
fcs/rcsi/hhs/hdds/fcm/fclcm add_x indicators}

\item{fsl_fcs_cereal}{the name of the variable that indicates the number of days cereals were consumed}

\item{fsl_fcs_legumes}{the name of the variable that indicates the number of days legumes were consumed}

\item{fsl_fcs_dairy}{the name of the variable that indicates the number of days dairy were consumed}

\item{fsl_fcs_meat}{the name of the variable that indicates the number of days meat were consumed}

\item{fsl_fcs_veg}{the name of the variable that indicates the number of days vegetables were consumed}

\item{fsl_fcs_fruit}{the name of the variable that indicates the number of days fruit were consumed}

\item{fsl_fcs_oil}{the name of the variable that indicates the number of days oil was consumed}

\item{fsl_fcs_sugar}{the name of the variable that indicates the number of days sugar was consumed}

\item{fsl_rcsi_lessquality}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to rely on less preferred and less expensive food to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_borrow}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to borrow food or rely on help from a relative or friend to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_mealsize}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to limit portion size of meals at meal times to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_mealadult}{Column representing question- During the last 7 days, were there days (and, if so, how many) when your household had to restrict consumption by adults in order for small children to eat to cope with a lack of food or money to buy it?}

\item{fsl_rcsi_mealnb}{Column representing question - During the last 7 days, were there days (and, if so, how many) when your household had to reduce number of meals eaten in a day to cope with a lack of food or money to buy it?}

\item{fsl_fcs_score}{Column representing FCS Score}

\item{fsl_rcsi_score}{Column representing rCSI Score}

\item{fsl_hhs_score}{Column representing HHS Score}

\item{fsl_hdds_score}{Column representing HDDS Score}

\item{uuid}{uuid variable}
}
\value{
a dataframe with all fsl related plausibility columns
}
\description{
create_muac_plaus
}
\examples{
\dontrun{create_muac_plaus(df)}
}
